Adds saving sample data (#36)
## 🎯 Aim

The aim of this PR is to add automotive support in gathering sample data
of ACES and SPFx extensions, libraries and webparts. The PR will add a
new workflow scheduled for Saturday to run a ps script that concats all
sample.json files from spfx-dev-fx-webpart/extension/component/aces pnp
repos and store them in a new data folder in vscode-viva repo (this
folder is not included in the extension file). Then those are used when
fetched to build sample views

## 📷 Result


![image](https://user-images.githubusercontent.com/58668583/236665094-416a6cd3-342f-4e6d-a2ba-0506248d0cad.png)

## ✅ What was done

- [X] added ps script to get all sample.json files form pnp sample repos
- [X] added a gh action workflow scheduled that will run the scripts
- [X] updated vscodeignore file

## 🔗 Related issue

Related: #6