Updates vscode to latest, small refactor and adds Copilot action to welcome view. Closes #334 (#392)
## 🎯 Aim

The aim is to update vscode types and approach with latest. Recheck and
cleanup what changed. When upgraded vscode we may use latest
improvements of how to handle model picking and loading history in chat
command. I checked that and unfrotonatly other models than gpt-4o bring
totally different responses which are not possible to pars with current
logic. Also loading history seems to be lacking some parts we get with
current implementation, so those parts are not changed.
Lastly I added action to activate SPFx Copilot from welcome view

## 📷 Result

![image](https://github.com/user-attachments/assets/8f676a95-62a2-4475-a946-b38ce2e2c811)

## ✅ What was done

- [X] updated vscode engine and typing to latest
- [X] refactor auth provider
- [X] rechecked current chat logic implementation
- [X] added copilot action to welcome view

## 🔗 Related Issue

Closes: #334