Increase security in GitHub Workflows. (#481)
## 🎯 Aim

The aim is to address security issues with workflows we used. 
Along the way I also added a condition so that the workflows will only
run in the context of our repo and not on the forked repos as well.
docs ->
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idif

Since GitHub workflows run based on repo default branch we should merge
this to `main` branch so it will resolve the security issues ASAP.

## ✅ What was done

- [X] Added permissions 
- [X] Added condition