Dependencies updates. Closes #434 (#464)
## 🎯 Aim

The aim of this PR is to update extension dependencies to the latest
version where possible. Unfortunately, as I checked, we are still not
able to use ESM version of the CLI for Microsoft 365, so this will need
to wait.
I also did not touch React version as I didn't want to cover to much.
Along the way I did some minor code cleanup and updated the `vsce`
package that we use in all of our workflows to current latest. I also
removed the package which we not use.
This PR may be a bit harder then the usual once as it would be best to
test it locally the same way you would install the VS Code extension
from package. In order to do that you will need to:

1. if you have SPFx Toolkit extension installed, then uninstall it
2. pull this branch and remove node_modules
3. run `npm install`
4. run `.\scripts\cli-for-microsoft365-copy-local-version.ps1
-workspacePath "<Your Path Here>"` script to get the cjs version of cli
from our fork
5. use `vsce package` to create local vscix package of the extension
6. in VS Code go to extension tab click o the 3 dots and install the
extension from file
7. after that it would be best to test basically everything also
checking the styles in the webviews

## ✅ What was done

- [X] updated `vsce` package in our workflows to latest
- [X] added a `qna` link in package.json to point to our discussion -
after this goes to `main` branch in VS Code market place when someone
will use the Q&A tab to ask questions that he or she will be redirected
to our discussion board in our repo
- [X] removed not used packages 
- [X] where possible updated the dev dependencies to latest except React
- [X] updated type packages to use exact version - from what I learned
we should always use exact version of type packages
- [X] updated `tailwindcss` to latest and updated the config to latest
and adapted the styles where it was needed
- [X] updated `typescript` to v5
- [X] updated to use exact versions in dependencies
- [X] minor code cleanup

## 🔗 Related issue

Closes: #434