Adds LM Tool to perform SPFx project upgrade. Closes #462 (#527)
## 🎯 Aim

The aim is to add and LM Tool that will not only allow to retrieve spfx
project upgrade guidance but also act uppon it and perform the project
upgrade. Further the LM is instructed to try to build the project and
check for errors in the terminal if something required further fixup.

### ⚠️ To Test this locally

In order to develop this feature I had to [tweak a bit our version of
CLI for Microsoft 365
](https://github.com/Adam-it/cli-microsoft365/commit/cb0949aa8be65b77ff925986582d0a3201a7bbe6)
exposing a new option `path` to allow to specify the path of the spfx
project which should be upgraded so in order to test it locally you need
to again follow the steps to build the cjs version of CLI for Microsoft
365 and run the script that updates the node modules of the extension
locally.
 
## 📷 Result


![spfx-upgrade1](https://github.com/user-attachments/assets/78fabfc0-3bd3-4341-889a-e7054aaa55d4)

![spfx-upgrade2](https://github.com/user-attachments/assets/3782556b-0b54-4023-8d08-d131ae77191f)

![spfx-upgrade3](https://github.com/user-attachments/assets/e10f8ad9-5f8c-4fe9-a21f-4602b9894d1d)

![spfx-upgrade4](https://github.com/user-attachments/assets/1d005ed0-80d4-46de-97e3-78dfc00b924a)

![spfx-upgrade5](https://github.com/user-attachments/assets/ee89c064-4b09-40a1-b9ab-0a7c4813325a)

![spfx-upgrade6](https://github.com/user-attachments/assets/fe29553f-c500-4cbe-afbf-fd7e3b696301)

![spfx-upgrade7](https://github.com/user-attachments/assets/7bd108d5-2ac9-41f3-9f3c-71b2c269ca21)

## ✅ What was done

- [X] Adds new LM Tool

## 🔗 Related issue

Closes: #462

---------

Co-authored-by: Saurabh Tripathi <saurabh7019@gmail.com>