Updates install dependencies and validate local setup to support all SPFx versions. Closes #516 #515 (#621)
The aim of this change is to refactor the install dependencies and
validate local setup actions so that they will now support all SPFx
versions.

Along the way, when validating node version and the user has in the SPFx
settings set to use either nvm or nvs as node package manager, we now
support switching to the correct version of node if present or
installing using either nvm or nvs (depending on what is selected in the
setting)

Validating node when NVM selected and option to install and switch to it

<img width="1919" height="988" alt="image"
src="https://github.com/user-attachments/assets/a64ac266-8541-4524-bb46-0ce83fba7b16"
/>

All SPFx versions picker

<img width="1431" height="844" alt="image"
src="https://github.com/user-attachments/assets/f7b963a8-58a2-4bbc-b77e-c2ef3b81e4ed"
/>

Validating node when NVS selected and option to install and switch to it

<img width="620" height="381" alt="image"
src="https://github.com/user-attachments/assets/4ff83942-be64-423e-9d4a-a185139a13d7"
/>

All validated properly - local environment is set up for the SPFx
version that is installed

<img width="1919" height="711" alt="image"
src="https://github.com/user-attachments/assets/9da3f8b7-9b1c-4f1b-953f-592b0e8580f1"
/>

`yo` is not installed warning

<img width="648" height="293" alt="image"
src="https://github.com/user-attachments/assets/c2587b19-7bb9-4dbb-8f87-7976bf7702a0"
/>

Installing dependencies for the selected SPFx version with wrong current
node version and resolving this using nvm

https://github.com/user-attachments/assets/61a4c27b-02a9-419a-9d21-85292fa356dd

validating local setup with wrong version of node for the currently
installed Spfx version

https://github.com/user-attachments/assets/b5bb565c-bad6-4cd8-b788-4ea74943e556

Some scenarios worth testing locally:

1. Validate local setup when no SPFx generator is installed - in this
case, SPFx Toolkit should suggest installing the latest, along with
dependencies. When installing the latest, it will also validate if the
current node version is valid for the latest version of SPFx
2. Validate local setup when some SPFx generator version is installed,
but the node is not valid - in this case, SPFx Toolkit should show node
is not valid and suggest setting up local env for that version of SPFx,
this will run install dependencies for the installed version of SPFx and
along the way validate node or install it if Node Version Manager is
used.
3. Validate local setup when some SPFx generator version is installed,
and node is ok, but other dependencies are not installed - in this case,
SPFx Toolkit should show a correct warning and suggest running install
dependencies for that version of SPFx
4. Validate local setup when all is good
5. Install dependencies when selecting SPFx version not supported on the
current Node - in this case SPFx Toolkit should validate the wrong node
version and suggest either downloading the correct one or it will switch
or install and switch to the correct version of Node if NVM or NVS is
set in SPFx Toolkit settings
6. Recheck installing Node using NVM on Windows
7. Recheck installing Node using NVM on Linux or Mac (we may use WSL on
Windows to validate this)
8. Recheck installing Node using NVS (this works the same on Windows,
Linux, and Mac)
9. Install dependencies - wait till the end of installation, and then
run validate and check if everything is fine (positive scenario)

- [x] added `src/constants/SpfxCompatibilityMatrix.ts` which now stores
all SPFx versions with the correct version of node(s) for each version
and dependencies, and release notes link which show up after someone
installs dependencies for this version. Also added (commented out) a
setup for SPFx 1.22. This is still commented out as 1.22 is in beta, so
the dependencies might still change.
- [x] implemented SPFx Doctor command to validate local user env setup
instead of our own. This may also be used in the future to validate the
local environment for SPFx version of the currently opened SPFx project.
Unfortunately, we may not use it to validate the correct version of node
- check the warning below so this is done by comparing with the
compatibility matrix.
- [x] added a `runCommandAndWait` in the `TerminalCommandExecuter.ts`
which allows executing a command in the terminal and waiting for it to
end. Needed so we are able to show that the local env is ready to go
when npm install ends.
- [x] added node validation based on `SpfxCompatibilityMatrix`.
- [x] if user sets in SPFx Toolkit that he or she uses either `nvs` or
`nvm` I added additional support that checks for the correct node
version and performs the switch or installs the correct node version
using the selected node version manager. This means we no longer point
to the download link, but now SPFx Toolkit is able to install Node for
you as well as long as you use either nvm or nvs.
- [x] refactored installing dependencies that now allow picking any
version of SPFx. After the SPFx version is selected, the extension will
first validate the node if it is compatible with the selected version.
If not it will either point to a download link or set up a node if Node
Version Manager is set in SPFx Toolkit settings. If the node was valid
or was corrected by SPFx Toolkit, then the extension proceeds with
installing the dependencies. It also now waits till the installation
ends successfully and presents a notification that all is set, with a
link to release notes for that specific SPFx version
- [x] updated documentation to align the changes

limitation

Please be aware that internally, VS Code is using Electron, which
bundles a specific version of Node. So when executing `spfx doctor`
within npm package and when it checks the Node version, it will not give
the Node version you are using on your PC, but the Node version used in
Electron used in VS Code. That is why we may not relly on node check
from `spfx doctor` command but will always need to do that ourselves by
executing `node -v` in user default terminal, checking for output and
validating with `SpfxCompatibilityMatrix`

Please be aware that NVM for Windows has a different set of commands and
output format than for linux and Mac. For example, on Windows we use
`nvm list` to check all node versions, and on Linux and Mac we use `nvm
ls`. That is why, when testing NVM support for installing node, we
should check both on Windows and Linux. We may use WSL for that

Closes #516
Closes #515

---------

Co-authored-by: Saurabh Tripathi <saurabh7019@gmail.com>