Updates astro package and moves images to different location. Closes #616 (#631)
## 🎯 Aim

The aim of this PR is to update the `astro` npm package which will also
update the sub dependencies and solve all security issues we have in
this repo.
As `astro` had an update to the way the images are handled, we also
agreed to move all images to `/dcos` subcatalog

## 🧪 How and what to test

We should check if images from new location render correctly so I
suggest checking the following:

- build and start the docs locally and recheck all pages if all images
are shown
- check images in extension by doing the following
    - if you have, uninstall SPFx Toolkit
    - build SPFx Toolkit vscode package locally by runing `vsce package`
    - install SPFx Toolkit from local file
    - recheck walkthrough
- recheck readme by showing the extension from the VS Code marketplace
inside VS Code - ⚠️ This is a bit tricky as the readme takes images
directly from repo so since this is not merged yet those will not
render. BUT what we may do is open the vscode webview developer tools
and manually check the paths (URLS) of each images in the readme and if
it is taken from the new location

## ✅ What was done

- [X] updated `astro` npm package
- [X] moved all images to `/docs/assets/images-vscode`
- [X] updated all paths in docs
- [X] updated all paths in package.json, readme, walkthrough
- [X] excluded `/docs/assets/images-vscode` from .vscodeignore file

## 🔗 Related issue

Closes: #616