Refactor the way we add CLI for Microsoft 365 to SPFx Toolkit. Closes #526 (#629)
The aim is to refactor SPFx Toolkit to use our own version of CLI for
Microsoft 365 npm package instead of the hacky current approach. As a
result, I published our own version of CLI for Microsoft 365 for SPFx
Toolkit
(https://www.npmjs.com/package/@pnp/cli-microsoft365-spfx-toolkit) in
CommonJS, and now SPFx Toolkit simply uses this package instead of doing
it manually. This simplified development also allows us to have a more
formal version of CLI for Microsoft 365 that is based on the official
package but has some tweaks created for this project

I would suggest to run `npm install` and `vsce package` to create the
local version of the extension locally and install it in VS Code and
click around the SPFx Toolkit mainly focusing on features that use CLI
for Microsoft 365 commands.

- [X] Updated npm package.json to use CLI for Microsoft 365 SPFx Toolkit
package and updated all imports
- [X] deleted the hacky script to ingest CLI for Microsoft 365 from the
local repo build
- [X] cleaned up contributing guidance
- [X] cleaned up GitHub Workflows not to build CLI for Microsoft 365
repo and added the files manually

Closes: #526